Fabric_assigment/sdk/contract/business_logic.go
package main

import (
	"fmt"

	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/fabric"
	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
)

type BusinessLogic struct{}

func (bl *BusinessLogic) CreateAccount(ctx fabsdk.TransactionContext, dealerID string, msisdn string, mpin string, balance int, status string, transAmount int, transType string, remarks string) error {
	fmt.Println("Creating account...")
	// Implement account creation logic here
	return nil
}

func (bl *BusinessLogic) UpdateAccount(ctx fabsdk.TransactionContext, dealerID string, balance int, transAmount int, transType string, remarks string) error {
	fmt.Println("Updating account...")
	// Implement account update logic here
	return nil
}

func (bl *BusinessLogic) GetAccount(ctx fabsdk.TransactionContext, dealerID string) ([]byte, error) {
	fmt.Println("Getting account...")
	// Implement account retrieval logic here
	return nil, nil
}

func (bl *BusinessLogic) GetHistory(ctx fabsdk.TransactionContext, dealerID string) ([]byte, error) {
	fmt.Println("Getting history...")
	// Implement history retrieval logic here
	return nil, nil
}